spiceroute/grocery-delivery-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
motor==3.3.1
pymongo==4.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
httpx==0.25.0
websockets==11.0.3
elasticsearch==8.10.0
python-dotenv==1.0.0
pydantic-settings==2.0.3 